Add generic cloud config upstream refresher
Refreshing and checking differences in upstream config spec for all
new cloud drivers can be done by generic service. Code base from existing
eksupstreamrefresher should be reused and prepared for adding new cloud drivers.

(cherry picked from commit 45c97f88ad39f263263a0d3b7fd9a51cd730bfa6)
