+++ pkg/controllers/management/wrangler.go
@@ -4,8 +4,8 @@ import (
 	"context"
 
 	"github.com/rancher/rancher/pkg/clustermanager"
+	"github.com/rancher/rancher/pkg/controllers/management/clusterupstreamrefresher"
 	"github.com/rancher/rancher/pkg/controllers/management/eks"
-	"github.com/rancher/rancher/pkg/controllers/management/eksupstreamrefresh"
 	"github.com/rancher/rancher/pkg/controllers/management/k3sbasedupgrade"
 	"github.com/rancher/rancher/pkg/controllers/management/systemcharts"
 	"github.com/rancher/rancher/pkg/types/config"
@@ -15,6 +15,6 @@ import (
 func RegisterWrangler(ctx context.Context, wranglerContext *wrangler.Context, management *config.ManagementContext, manager *clustermanager.Manager) error {
 	k3sbasedupgrade.Register(ctx, wranglerContext, management, manager)
 	eks.Register(ctx, wranglerContext, management)
-	eksupstreamrefresh.Register(ctx, wranglerContext)
+	clusterupstreamrefresher.Register(ctx, wranglerContext)
 	return systemcharts.Register(ctx, wranglerContext)
 }
